requirements.txt
langchain
langchain-core
langchain-google-genai
langchain-community
google-generativeai
faiss-cpu
pypdf
python-dotenv
sentence-transformers
streamlit